+++ src/DataHelpers.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.15;
 
-function packTokenId(uint challengeId, uint32 level) pure returns (uint) {
+function packTokenId(uint256 challengeId, uint32 level) pure returns (uint256) {
     return (challengeId << 32) | level;
 }
