+++ src/TokenDetails.sol
@@ -4,15 +4,13 @@ pragma solidity ^0.8.15;
 import {IChallenge} from "./IChallenge.sol";
 
 struct TokenDetails {
-    uint challengeId;
+    uint256 challengeId;
     IChallenge challenge;
-
     uint32 leaderGas;
     uint32 leaderLevel;
     address leaderSolver;
     address leaderOwner;
     address leaderSubmission;
-
     uint32 gas;
     uint32 level;
     uint32 rank;
